Rename `subscribe-with-google` folder to `thank-with-google`.
