+++ assets/js/modules/thank-with-google/components/common/PublicationIDInput.js
