+++ assets/js/modules/thank-with-google/components/common/index.js
