+++ assets/js/modules/thank-with-google/components/settings/SettingsEdit.js
