+++ assets/js/modules/thank-with-google/components/settings/SettingsForm.js
