+++ assets/js/modules/thank-with-google/components/settings/SettingsView.js
