+++ assets/js/modules/thank-with-google/components/settings/index.js
