+++ assets/js/modules/thank-with-google/components/setup/SetupMain.js
