+++ assets/js/modules/thank-with-google/components/setup/index.js
