+++ assets/js/modules/thank-with-google/datastore/settings.js
