+++ assets/js/modules/thank-with-google/util/validation.js
